BUG: Typo
git-svn-id: http://svn.slicer.org/Slicer4@5824 3bd1e089-480b-0410-8dfb-8563597acbee

+++ branches/igt_cuda_merge/Modules/VolumeRenderingCuda/cudaTypeRange.h
@@ -35,7 +35,7 @@ void GetTypeRange<unsigned short>(unsigned short& min, unsigned short& max)
 }
 
 template<>
-void GetTypeRange<int>(int short& min, int& max)
+void GetTypeRange<int>(int& min, int& max)
 {
     min = INT_MIN;
     max = INT_MAX;
